+++ src/main/java/com/cibertec/aplicacionesmovilesiiantony/service/abstraccion/UserService.java
@@ -1,5 +1,6 @@
 package com.cibertec.aplicacionesmovilesiiantony.service.abstraccion;
 
+import com.cibertec.aplicacionesmovilesiiantony.model.Rol;
 import com.cibertec.aplicacionesmovilesiiantony.model.User;
 import org.springframework.http.ResponseEntity;
 
@@ -12,7 +13,7 @@ public interface UserService {
     ResponseEntity<String> signUp(Map<String, String> requestMap);
     ResponseEntity<?> updatePassword(Map<String, String> requestMap);
     ResponseEntity<?> updateUser(Map<String, String> requestMap);
-    User findById(Long id);
-    void deleteForId(Long id);
+    User findById(Integer id);
+    void deleteForId(Integer id);
 
 }
